Add fill-column-indicator
There is no direct evidence for the placement of this commit before
the next three, as no copies of the .emacs file with this change
but not the next three are known to survive. It is only known that
fill-column-indicator.el was added at the time listed in this commit.

However, analyzing the commit history of the CS168 routing project, one
can see a bunch of commits on the days leading up to October 7. It
is most probable that the recentf and window scrolling stuff were added
during those days, as:
 - I would have gotten tired of typing dv_router.py over and over again
 - dv_router.py was a long file and I had to frequently scroll back and
   forth to cross-reference function invocations with their
   declarations; the impetus for adding a means to scroll the window by
   a small amount and for not wrecking the cursor position would have
   been large at this time.